requirements.txt
# Core Packages
Django==5.1.7
djangorestframework==3.15.0
django-cors-headers==4.3.0
psycopg2-binary==2.9.9
gunicorn==21.2.0
whitenoise==6.9.0
python-decouple==3.8
django-environ==0.12.0

# Development & Testing
django-debug-toolbar==4.0.0
pytest==8.0.0
pytest-django==4.11.1
coverage==7.0.0
ipython==8.0.0

# Security & Utilities
Pillow==10.3.0
asgiref==3.8.1
sqlparse==0.5.3